Split controller manager options from init
Make defaults and flags clearly distinct from initialization code.
