+++ cmd/hyperkube/kube-controller-manager.go
@@ -19,19 +19,20 @@ limitations under the License.
 package main
 
 import (
-	controllermgr "k8s.io/kubernetes/cmd/kube-controller-manager/app"
+	"k8s.io/kubernetes/cmd/kube-controller-manager/app"
+	"k8s.io/kubernetes/cmd/kube-controller-manager/app/options"
 )
 
 // NewKubeControllerManager creates a new hyperkube Server object that includes the
 // description and flags.
 func NewKubeControllerManager() *Server {
-	s := controllermgr.NewCMServer()
+	s := options.NewCMServer()
 
 	hks := Server{
 		SimpleUsage: "controller-manager",
 		Long:        "A server that runs a set of active components. This includes replication controllers, service endpoints and nodes.",
 		Run: func(_ *Server, args []string) error {
-			return s.Run(args)
+			return app.Run(s)
 		},
 	}
 	s.AddFlags(hks.Flags())
